+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - [feat:support traffic mirroring.](https://github.com/Tencent/spring-cloud-tencent/pull/1687)
 - [feat: support custom quickstart circuitbreak delay time.](https://github.com/Tencent/spring-cloud-tencent/pull/1688)
 - [feat: add delay interface in tsf-example.](https://github.com/Tencent/spring-cloud-tencent/pull/1689)
+- [fix: fix lb configuration on bootstrap step.](https://github.com/Tencent/spring-cloud-tencent/issues/1690)